requirements.txt
attrs==22.2.0
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.1
cssselect==1.2.0
filelock==3.10.7
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
json5==0.9.11
lxml==4.9.2
packaging==23.0
parsel==1.7.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.1.1
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
scrapeops-scrapy==0.5.2
scrapeops-scrapy-proxy-sdk==1.0
Scrapy==2.8.0
service-identity==21.1.0
six==1.16.0
tld==0.13
tldextract==3.4.0
Twisted==22.10.0
typing_extensions==4.5.0
urllib3==1.26.15
w3lib==2.1.1
zope.interface==6.0
python-dotenv==1.0.0